requirements.txt
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
h11==0.14.0
idna==3.10
mutagen==1.47.0
outcome==1.3.0.post0
packaging==24.2
pycparser==2.22
pydub==0.25.1
PySocks==1.7.1
python-dotenv==1.0.1
pytube==15.0.0
redis==5.2.1
requests==2.32.3
selenium==4.27.1
setuptools==75.8.0
sniffio==1.3.1
sortedcontainers==2.4.0
spotipy==2.24.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yt-dlp==2025.1.15
ytmusicapi==1.8.2
